Remove the protobuf::__runtime::Map type
Before this change the runtimes export a Map and MapInner. This change merges the Map and MapInner types into a single MapInner type, removing the Map type from the runtime.

The motivation for this change is that a separate Map thpe is not needed by the runtime and so that we can create a public runtime-agnostic protobuf::Map type that implements Proxied. This will be done in a follow-up CL.

PiperOrigin-RevId: 581944445